requirements.txt
ipykernel
python-dotenv
langchain
langchain-ollama
langgraph
langchain_community
tavily-python
langgraph-checkpoint-sqlite